+++ src/Rhetos.LanguageServices.Server/Handlers/RhetosCompletionHandler.cs
@@ -38,6 +38,8 @@ namespace Rhetos.LanguageServices.Server.Handlers
         {
             DocumentSelector = TextDocumentHandler.RhetosDocumentSelector,
             ResolveProvider = true,
+            TriggerCharacters = new Container<string>("."),
+            AllCommitCharacters = new Container<string>(" ", ".")
         };
 
         public RhetosCompletionHandler(RhetosWorkspace rhetosWorkspace, ConceptQueries conceptQueries, ILogger<RhetosCompletionHandler> log)