+++ src/Rhetos.LanguageServices.Server/Services/RhetosProjectMonitor.cs
@@ -1,0 +1,86 @@
+﻿/*
+    Copyright (C) 2014 Omega software d.o.o.
+
+    This file is part of Rhetos.
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as
+    published by the Free Software Foundation, either version 3 of the
+    License, or (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+using System;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+
+namespace Rhetos.LanguageServices.Server.Services
+{
+    public class RhetosProjectMonitor
+    {
+        private static readonly TimeSpan _cycleInterval = TimeSpan.FromMilliseconds(1000);
+        private readonly RhetosWorkspace rhetosWorkspace;
+        private readonly ILogger<RhetosProjectMonitor> log;
+
+        private readonly CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
+        private Task monitorLoopTask;
+
+        public RhetosProjectMonitor(RhetosWorkspace rhetosWorkspace, ILogger<RhetosProjectMonitor> log)
+        {
+            this.rhetosWorkspace = rhetosWorkspace;
+            this.log = log;
+        }
+
+        public void Start()
+        {
+            if (cancellationTokenSource.IsCancellationRequested)
+                return;
+
+            log.LogInformation($"Starting {nameof(RhetosProjectMonitor)}.{nameof(MonitorLoop)}.");
+            monitorLoopTask = Task.Factory.StartNew(() => MonitorLoop(cancellationTokenSource.Token), TaskCreationOptions.LongRunning);
+        }
+
+        public void Stop()
+        {
+            try
+            {
+                log.LogDebug($"Stopping {nameof(RhetosProjectMonitor)}.{nameof(MonitorLoop)}.");
+                cancellationTokenSource.Cancel();
+                monitorLoopTask?.Wait();
+            }
+            catch (Exception e)
+            {
+                if (e is AggregateException aggregateException && aggregateException.InnerExceptions.Any(inner => !(inner is TaskCanceledException)))
+                    log.LogDebug($"{nameof(MonitorLoop)} successfully cancelled.");
+                else
+                    log.LogDebug($"{nameof(MonitorLoop)} faulted while waiting to cancel: {monitorLoopTask?.Exception}");
+            }
+        }
+
+        public void MonitorLoop(CancellationToken cancellationToken)
+        {
+            while (true)
+            {
+                Task.Delay(_cycleInterval, cancellationToken).Wait(cancellationToken);
+
+                try
+                {
+                    rhetosWorkspace.UpdateRhetosContextStatus();
+                }
+                catch (Exception e)
+                {
+                    log.LogWarning($"Error occured during monitoring RhetosAppContext state: {e}");
+                }
+            }
+        }
+    }
+}